Move CPossessionDrawConfiguration traits
