+++ include/football/CPossessionDrawConfiguration.h
@@ -1,38 +1,12 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/default.h"
-#include "traits/json.h"
+#include "football/traits/CPossessionDrawConfiguration.h"
 
 #include "football/types/CDrawConfiguration.h"
 #include "football/types/CPossessionDrawConfiguration.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CPossessionDrawConfiguration;
-}
-
-template <> struct default_traits<football::CPossessionDrawConfiguration>
-{
-protected:
-	using probability = football::types::CDrawConfiguration::probability;
-public:
-	//! Default keep posssession probability.
-	static inline constexpr probability KEEP_POSSESSION_PROBABILITY = probability{ 295952 } / 460939;
-};
-
-template <> struct json_traits<football::CPossessionDrawConfiguration>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Possession draw configuration";
-	//! JSON key for the \copybrief football::CPossessionDrawConfiguration::mKeepPossessionProbability
-	static inline constexpr std::string_view KEEP_POSSESSION_PROBABILITY_KEY = "Keep possession probability";
-};
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -72,6 +46,4 @@ private:
 	probability mKeepPossessionProbability;
 };
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
